agno_db_server/requirements.txt
psycopg[binary]
psycopg2
mysqlclient
sqlalchemy
agno
python-dotenv
requests
openai
fastapi[standard]
duckdb